Wheezy: Fix build issues with Docker mk-builder image
The Travis Docker mk-build images for Wheezy armhf builds use a
cross compiler for speed.  Due to assorted bad luck:

http://stackoverflow.com/a/27944574

...the gcc-4.7 cross compiler for Wheezy is not readily available,
so the 4.9 version from Jessie is used for Travis builds (which
must run in a Docker instance on amd64 hardware).  This commit
adds a "wheezy_hacks.c" file which provids the missing ABI
functions expected by gcc-4.9 but not present in libstdc++ 4.7
used when building Machinekit on Wheezy and updates the gcc/g++
wrapper scripts to add the "wheezy_hacks.o" object file to the
command line whenever -lstdc++ is present in the gcc/g++ argument
list.

Signed-off-by: Charles Steinkuehler <charles@steinkuehler.net>